Move Parse* functions to sig_helpers.cpp
