+++ src/Datadog.Trace.ClrProfiler.Native/sig_helpers.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <corhlpr.h>
+
+namespace trace {
+
+bool ParseType(PCCOR_SIGNATURE* p_sig);
+
+}  // namespace trace
